Fix: Id type changed to Long

+++ manage/src/main/java/com/system/manage/controllers/professorController.java
@@ -27,7 +27,7 @@ public class professorController {
     }
 
     @PostMapping("professor_form")
-    public ModelAndView cadastro_professor(@RequestParam("id")Integer id, @RequestParam("nome") String nome,
+    public ModelAndView cadastro_professor(@RequestParam("id")Long id, @RequestParam("nome") String nome,
     @RequestParam("academic") String formacao, @RequestParam("list") String  disc, @RequestParam("bool") Boolean status) {
         ModelAndView mv = new ModelAndView();
         mv.setViewName("home/index");
@@ -42,7 +42,7 @@ public class professorController {
             savior.setId(id);
             savior.setNome(nome);
             savior.setAcademicalInfo(formacao);
-            savior.setCode(Integer.toString(id));
+            savior.setCode(Long.toString(id));
             String[] new_disc = disc.split(";");
             for (int i = 0; i < new_disc.length; i++) {
                 savior.setList(new_disc[i]);
@@ -73,7 +73,7 @@ public class professorController {
     }
 
     @GetMapping("/alterar/{id}")
-    public ModelAndView alterar(@PathVariable("id") Integer id) {
+    public ModelAndView alterar(@PathVariable("id") Long id) {
         ModelAndView mv = new ModelAndView();
         try {
             mv.setViewName("forms/alterar_prof");
@@ -103,7 +103,7 @@ public class professorController {
     }
 
     @GetMapping("/remover/{id}")
-    public ModelAndView excluirProfessor(@PathVariable("id") Integer id) {
+    public ModelAndView excluirProfessor(@PathVariable("id") Long id) {
         ModelAndView mv = new ModelAndView();
         try {
             repo.deleteById(id);
